做完了大部分功能，mod未完全测试，现在可以做到伤害修改有效且不报错

+++ src/main/java/moe/gensoukyo/rpgmaths/api/impl/stats/StatHandlerImpl.java
@@ -6,32 +6,29 @@ import net.minecraftforge.common.capabilities.ICapabilityProvider;
 
 /**
  * StatHandler的实现
+ * 与实体绑定
  * @author Chloe_koopa
  * @see IStatHandler
  */
-public class StatHandlerImpl implements IStatHandler
-{
+public class StatHandlerImpl implements IStatHandler {
     protected ICapabilityProvider owner;
-    public StatHandlerImpl(ICapabilityProvider owner)
-    {
+
+    public StatHandlerImpl(ICapabilityProvider owner) {
         this.owner = owner;
     }
 
     @Override
-    public float getBaseValue(IStatType type)
-    {
+    public double getBaseValue(IStatType type) {
         return type.getBaseValue(owner);
     }
 
     @Override
-    public float getFinalValue(IStatType type)
-    {
+    public double getFinalValue(IStatType type) {
         return type.getFinalValue(owner);
     }
 
     @Override
-    public boolean setBaseValue(IStatType type, float value)
-    {
+    public boolean setBaseValue(IStatType type, double value) {
         return type.setBaseValue(owner, value);
     }
 }
